refactoring, add yajra package, emove comment, new method implement yajra datatables

See merge request energeek/universitas-pertamina-api!52

+++ app/Http/Controllers/ProfileController.php
@@ -32,7 +32,6 @@ class ProfileController extends Controller
            $this->responseData = new ProfileInstitutionDataResource($data);
         } else if ($user->type == 1) {
             $data = Member::with('title')->with('memberSkill')->with('memberResearchInterest')->with('memberEducation')->with('department')->with('nationality')->with('publication')->find($user->owner_id);
-            // $this->responseData = $data;
             $this->responseData = new ProfileMemberDataResource($data);
         } else {
             $this->responseData = $user;
@@ -62,7 +61,6 @@ class ProfileController extends Controller
      public function storeInstitution(ProfileInstitutionStoreRequest $request, Institution $institution)
      {
          $request->validated();
-         // $institution->email = $request->input('email');
          $institution->country = $request->input('country');
          $institution->city = $request->input('city');
          $institution->address = $request->input('address');
@@ -72,7 +70,7 @@ class ProfileController extends Controller
 
          //Publication//
          $department = $request->input('department');
-         $memberEducation = Department::where('institution_id', $institution->id)->delete();
+         Department::where('institution_id', $institution->id)->delete();
 
          foreach ($department as $key => $value) {
             Department::withTrashed()->updateOrCreate(
@@ -86,17 +84,15 @@ class ProfileController extends Controller
          //////////////////////////////////////////////////
 
          $file = $request->file('photo');
-         if (!empty($file)) {
-            if ($file->isValid()) {
-               $changedName = time().rand(100,999).$file->getClientOriginalName();
-               $file->storeAs('profile/institution/' . $institution->id, $changedName);
+         if (!empty($file) && $file->isValid()) {
+            $changedName = time().random_int(100,999).$file->getClientOriginalName();
+            $file->storeAs('profile/institution/' . $institution->id, $changedName);
 
-               if ($institution->path_photo != ''){
-                  unlink(storage_path('app/profile/institution/').$institution->id.'/'.$institution->path_photo);
-               }
-
-               $institution->path_photo = $changedName;
+            if ($institution->path_photo != ''){
+                unlink(storage_path('app/profile/institution/').$institution->id.'/'.$institution->path_photo);
             }
+
+            $institution->path_photo = $changedName;
          }
 
          $institution->save();
@@ -113,7 +109,6 @@ class ProfileController extends Controller
          $request->validated();
 
          $member->name = $request->input('name');
-         // $member->email = $request->input('email');
          $member->desc = $request->input('desc');
          $member->department_id = $request->input('department');
          $member->position = $request->input('position');
@@ -170,17 +165,15 @@ class ProfileController extends Controller
 
          //PHOTO//
          $file = $request->file('photo');
-         if (!empty($file)) {
-            if ($file->isValid()) {
-               $changedName = time().rand(100,999).$file->getClientOriginalName();
-               $file->storeAs('profile/member/' . $member->id, $changedName);
+         if (!empty($file) && $file->isValid()) {
+            $changedName = time().random_int(100,999).$file->getClientOriginalName();
+            $file->storeAs('profile/member/' . $member->id, $changedName);
 
-               if ($member->path_photo != ''){
-                  unlink(storage_path('app/profile/member/').$member->id.'/'.$member->path_photo);
-               }
-
-               $member->path_photo = $changedName;
+            if ($member->path_photo != ''){
+                unlink(storage_path('app/profile/member/').$member->id.'/'.$member->path_photo);
             }
+
+            $member->path_photo = $changedName;
          }
 
          $member->save();
